# Merge all embedding lookups (Cleaner but 1,0006448126x slower)