Currently embedding lookup is done 4 different times (1x input embeddings, 3x Value Embeddings). This can be done as
```
embeds = self.embed(input_seq).chunk(4, dim=-1)
```
Which is mathematically equivalent. After timing it it seems like the compiler must have realized this as well. The new change seem to be run identical to 10ms/step slower per step but arguably conceptually cleaner. 
Clean code tends to be easier to reason and find optimizations for so it's a matter of taste if it's worth it at the expense of a `(186988+187005)/(186882+186870) = 1,0006448126x` slower run.

Running with
```
torch.manual_seed(42)
torch.cuda.manual_seed(42)
torch.cuda.manual_seed_all(42)
torch.backends.cudnn.deterministic = True
torch.backends.cudnn.benchmark = True
```
master
```python
step:1393/1393 val_loss:3.2769 train_time:186882ms step_avg:135.13ms
peak memory allocated: 37653 MiB reserved: 41736 MiB

step:1393/1393 val_loss:3.2765 train_time:186870ms step_avg:135.12ms
peak memory allocated: 37653 MiB reserved: 41736 MiB
```

This branch
```python
step:1393/1393 val_loss:3.2805 train_time:186988ms step_avg:135.20ms
peak memory allocated: 37653 MiB reserved: 41656 MiB

step:1393/1393 val_loss:3.2797 train_time:187005ms step_avg:135.22ms
peak memory allocated: 37653 MiB reserved: 38856 MiB
```