@KellerJordan 2025-02-02
thanks, a variant of this has now been put into the code, with a reference to this PR