@YouJiacheng 2025-03-28
For context, this actually is the earliest implementation.
At that time, we used DDP, and single embedding + chunk is slower than split embeddings.